pip.requirements.txt
-e git://github.com/django/django.git@43510cffcbf64e73232da1d6209151c7232fd2b6#egg=Django-dev
djangorestframework==2.3.13
